Renamed TabStateTabContainerPanel to TabStateContainerPanel
