--- client-app/src/desktop/tabs/containers/tab/createContainerModelConfig.js
@@ -1,27 +1,0 @@
-import {div} from '@xh/hoist/cmp/layout';
-import {Icon} from '@xh/hoist/icon';
-
-export const createContainerModelConfig = (args) => {
-    const tabTxt = title => div(`This is the ${title} tab`);
-
-    return {
-        tabs: [
-            {
-                id: 'people',
-                icon: Icon.user(),
-                content: () => tabTxt('People')
-            },
-            {
-                id: 'places',
-                icon: Icon.home(),
-                content: () => tabTxt('Places')
-            },
-            {
-                id: 'things',
-                icon: Icon.portfolio(),
-                content: () => tabTxt('Things')
-            }
-        ],
-        ...args
-    };
-};
